x86/tpm.c: hashing and extending PCRs for TPM1.2
This file is built twice: for early 32b mode without paging and for 64b
code.  The expectation is that the data that's measured early is small
and thus sending it to TPM to do the hashing is viable.  Version with
paging computes digests and only sends their values to TPM, thus
permitting hashing of large chunks of data like dom0's kernel and
initrd (sending them to TPM would take multiple minutes).

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>
Signed-off-by: Sergii Dmytruk <sergii.dmytruk@3mdeb.com>