x86: Activate Data Operand Invariant Timing Mode by default
Intel IceLake and later CPUs have microarchitectural behaviors which cause
data-dependent timing behavior.  This is not an issue for 99% of software,
but it is a problem for cryptography routines.  On these CPUs, a new
architectural feature, DOITM, was retrofitted in microcode.

For now, Xen can't enumerate DOITM to guest kernels; getting this working is
still in progress.  The consequence is that guest kernels will incorrectly
conclude that they are safe.

To maintain the safety of current software, activate DOITM unilaterally.  This
will be relaxed in the future when we can enumerate the feature properly to
guests.

[ Demi: add missing MSR values and drop a stopgap command-line option ]

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Co-authored-by: Demi Marie Obenour <demi@invisiblethingslab.com>
Signed-off-by: Demi Marie Obenour <demi@invisiblethingslab.com>