x86/boot: choose AP stack based on APIC ID
This is made as first step of making parallel AP bring-up possible.
It should be enough for pre-C code.

CPU id is obtained as a side effect of searching for appropriate
stack for AP. It can be used as a parameter to start_secondary(),
instead of repeating the lookup done in assembly.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>