tpm: fixing up the time out loops
The tboot environment did not have an accurate way of tracking time outs
on access to the TPM hardware. As such, it relied on counting loops and
using a multiplier factor to estimate amount of time a loop would take.
Now that the code is under Xen, it can use the time functions to
determine an accurate(ly) time out.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>